+++ wp-content/themes/automize/woocommerce/content-quick-view-product.php
@@ -7,13 +7,29 @@
 
 global $product;
 use RisingBambooTheme\Helper\Setting;
-$outside         = Setting::get(RISING_BAMBOO_KIRKI_FIELD_COMPONENT_MODAL_CLICK_OUTSIDE_CLOSE);
-$modal_effect    = Setting::get(RISING_BAMBOO_KIRKI_FIELD_COMPONENT_MODAL_EFFECT);
-$backdrop_filter = Setting::get(RISING_BAMBOO_KIRKI_FIELD_COMPONENT_MODAL_BACKDROP_FILTER);
-$classes         = [ 'quick-view-product-wrap' ];
-$classes[]       = ( true === $backdrop_filter ) ? 'backdrop' : 'backdrop-none';
-$classes[]       = ( false === $outside ) ? 'outside-modal' : '';
-$class_string    = implode(' ', array_filter($classes));
+$outside           = Setting::get(RISING_BAMBOO_KIRKI_FIELD_COMPONENT_MODAL_CLICK_OUTSIDE_CLOSE);
+$modal_effect      = Setting::get(RISING_BAMBOO_KIRKI_FIELD_COMPONENT_MODAL_EFFECT);
+$backdrop_filter   = Setting::get(RISING_BAMBOO_KIRKI_FIELD_COMPONENT_MODAL_BACKDROP_FILTER);
+$show_excerpt      = Setting::get(RISING_BAMBOO_KIRKI_FIELD_WOOCOMMERCE_PRODUCT_DETAIL_SUMMARY_SHOW_EXCERPT);
+$classes           = [ 'quick-view-product-wrap' ];
+$classes[]         = ( true === $backdrop_filter ) ? 'backdrop' : 'backdrop-none';
+$classes[]         = ( false === $outside ) ? 'outside-modal' : '';
+$class_string      = implode(' ', array_filter($classes));
+$short_description = apply_filters('woocommerce_short_description', $product->get_short_description());
+if ( $show_excerpt && ! empty($short_description) ) {
+	add_action(
+		'woocommerce_single_product_summary',
+		function() use ( $product ) {
+			$short_description = apply_filters('woocommerce_short_description', $product->get_short_description());
+			if ( ! empty($short_description) ) {
+				echo '<div class="woocommerce-product-details__short-description mt-2 text-sm leading-relaxed">';
+				echo wp_kses($short_description, 'rbb-kses');
+				echo '</div>';
+			}
+		},
+		15
+	);
+}
 ?>
 
 <div class="<?php echo esc_attr($class_string); ?>" id="quick-view-product-wrap-<?php echo esc_attr($product->get_id()); ?>" data-modal-animation="<?php echo esc_attr($modal_effect); ?>">
@@ -40,7 +56,7 @@ $class_string    = implode(' ', array_filter($classes));
 					?>
 				</div>
 				<div class="product-detail-info flex relative overflow-hidden flex-col">
-					<div class="content_product_detail entry-summary w-full h-full absolute overflow-y-auto px-5 lg:pr-[30px] left-0">
+					<div class="content_product_detail text-left entry-summary w-full h-full absolute overflow-y-auto px-5 lg:pr-[30px] left-0">
 						<?php
 						/**
 						 * Hook woocommerce_single_product_summary