+++ wp-content/themes/automize/woocommerce/order/order-details.php
@@ -12,7 +12,7 @@
  *
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 9.6.0
+ * @version 10.1.0
  */
 
 use RisingBambooTheme\App\App;