+++ wp-content/themes/automize/woocommerce/single-product/tabs/tabs.php
@@ -12,7 +12,7 @@
  *
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 9.6.0
+ * @version 9.8.0
  */
 
 use RisingBambooTheme\Helper\Setting;